+++ vpx_dsp/x86/highbd_idct16x16_add_sse2.c
@@ -56,20 +56,20 @@ static INLINE void highbd_idct16_4col(__m128i *const io /*io[16]*/) {
   __m128i step1[16], step2[16];
 
   // stage 2
-  highbd_butterfly_sse2(io[1], io[15], (int)cospi_30_64, (int)cospi_2_64,
-                        &step2[8], &step2[15]);
-  highbd_butterfly_sse2(io[9], io[7], (int)cospi_14_64, (int)cospi_18_64,
-                        &step2[9], &step2[14]);
-  highbd_butterfly_sse2(io[5], io[11], (int)cospi_22_64, (int)cospi_10_64,
-                        &step2[10], &step2[13]);
-  highbd_butterfly_sse2(io[13], io[3], (int)cospi_6_64, (int)cospi_26_64,
-                        &step2[11], &step2[12]);
+  highbd_butterfly_sse2(io[1], io[15], cospi_30_64, cospi_2_64, &step2[8],
+                        &step2[15]);
+  highbd_butterfly_sse2(io[9], io[7], cospi_14_64, cospi_18_64, &step2[9],
+                        &step2[14]);
+  highbd_butterfly_sse2(io[5], io[11], cospi_22_64, cospi_10_64, &step2[10],
+                        &step2[13]);
+  highbd_butterfly_sse2(io[13], io[3], cospi_6_64, cospi_26_64, &step2[11],
+                        &step2[12]);
 
   // stage 3
-  highbd_butterfly_sse2(io[2], io[14], (int)cospi_28_64, (int)cospi_4_64,
-                        &step1[4], &step1[7]);
-  highbd_butterfly_sse2(io[10], io[6], (int)cospi_12_64, (int)cospi_20_64,
-                        &step1[5], &step1[6]);
+  highbd_butterfly_sse2(io[2], io[14], cospi_28_64, cospi_4_64, &step1[4],
+                        &step1[7]);
+  highbd_butterfly_sse2(io[10], io[6], cospi_12_64, cospi_20_64, &step1[5],
+                        &step1[6]);
   step1[8] = _mm_add_epi32(step2[8], step2[9]);
   step1[9] = _mm_sub_epi32(step2[8], step2[9]);
   step1[10] = _mm_sub_epi32(step2[10], step2[11]);  // step1[10] = -step1[10]
@@ -81,11 +81,11 @@ static INLINE void highbd_idct16_4col(__m128i *const io /*io[16]*/) {
 
   // stage 4
   highbd_butterfly_cospi16_sse2(io[0], io[8], &step2[0], &step2[1]);
-  highbd_butterfly_sse2(io[4], io[12], (int)cospi_24_64, (int)cospi_8_64,
-                        &step2[2], &step2[3]);
-  highbd_butterfly_sse2(step1[14], step1[9], (int)cospi_24_64, (int)cospi_8_64,
-                        &step2[9], &step2[14]);
-  highbd_butterfly_sse2(step1[10], step1[13], (int)cospi_8_64, (int)cospi_24_64,
+  highbd_butterfly_sse2(io[4], io[12], cospi_24_64, cospi_8_64, &step2[2],
+                        &step2[3]);
+  highbd_butterfly_sse2(step1[14], step1[9], cospi_24_64, cospi_8_64, &step2[9],
+                        &step2[14]);
+  highbd_butterfly_sse2(step1[10], step1[13], cospi_8_64, cospi_24_64,
                         &step2[13], &step2[10]);
   step2[5] = _mm_sub_epi32(step1[4], step1[5]);
   step1[4] = _mm_add_epi32(step1[4], step1[5]);
@@ -106,20 +106,20 @@ static INLINE void highbd_idct16x16_38_4col(__m128i *const io /*io[16]*/) {
   __m128i temp1[2], sign[2];
 
   // stage 2
-  highbd_partial_butterfly_sse2(io[1], (int)cospi_30_64, (int)cospi_2_64,
-                                &step2[8], &step2[15]);
-  highbd_partial_butterfly_neg_sse2(io[7], (int)cospi_14_64, (int)cospi_18_64,
-                                    &step2[9], &step2[14]);
-  highbd_partial_butterfly_sse2(io[5], (int)cospi_22_64, (int)cospi_10_64,
-                                &step2[10], &step2[13]);
-  highbd_partial_butterfly_neg_sse2(io[3], (int)cospi_6_64, (int)cospi_26_64,
-                                    &step2[11], &step2[12]);
+  highbd_partial_butterfly_sse2(io[1], cospi_30_64, cospi_2_64, &step2[8],
+                                &step2[15]);
+  highbd_partial_butterfly_neg_sse2(io[7], cospi_14_64, cospi_18_64, &step2[9],
+                                    &step2[14]);
+  highbd_partial_butterfly_sse2(io[5], cospi_22_64, cospi_10_64, &step2[10],
+                                &step2[13]);
+  highbd_partial_butterfly_neg_sse2(io[3], cospi_6_64, cospi_26_64, &step2[11],
+                                    &step2[12]);
 
   // stage 3
-  highbd_partial_butterfly_sse2(io[2], (int)cospi_28_64, (int)cospi_4_64,
-                                &step1[4], &step1[7]);
-  highbd_partial_butterfly_neg_sse2(io[6], (int)cospi_12_64, (int)cospi_20_64,
-                                    &step1[5], &step1[6]);
+  highbd_partial_butterfly_sse2(io[2], cospi_28_64, cospi_4_64, &step1[4],
+                                &step1[7]);
+  highbd_partial_butterfly_neg_sse2(io[6], cospi_12_64, cospi_20_64, &step1[5],
+                                    &step1[6]);
   step1[8] = _mm_add_epi32(step2[8], step2[9]);
   step1[9] = _mm_sub_epi32(step2[8], step2[9]);
   step1[10] = _mm_sub_epi32(step2[10], step2[11]);  // step1[10] = -step1[10]
@@ -131,13 +131,13 @@ static INLINE void highbd_idct16x16_38_4col(__m128i *const io /*io[16]*/) {
 
   // stage 4
   abs_extend_64bit_sse2(io[0], temp1, sign);
-  step2[0] = multiplication_round_shift_sse2(temp1, sign, (int)cospi_16_64);
+  step2[0] = multiplication_round_shift_sse2(temp1, sign, cospi_16_64);
   step2[1] = step2[0];
-  highbd_partial_butterfly_sse2(io[4], (int)cospi_24_64, (int)cospi_8_64,
-                                &step2[2], &step2[3]);
-  highbd_butterfly_sse2(step1[14], step1[9], (int)cospi_24_64, (int)cospi_8_64,
-                        &step2[9], &step2[14]);
-  highbd_butterfly_sse2(step1[10], step1[13], (int)cospi_8_64, (int)cospi_24_64,
+  highbd_partial_butterfly_sse2(io[4], cospi_24_64, cospi_8_64, &step2[2],
+                                &step2[3]);
+  highbd_butterfly_sse2(step1[14], step1[9], cospi_24_64, cospi_8_64, &step2[9],
+                        &step2[14]);
+  highbd_butterfly_sse2(step1[10], step1[13], cospi_8_64, cospi_24_64,
                         &step2[13], &step2[10]);
   step2[5] = _mm_sub_epi32(step1[4], step1[5]);
   step1[4] = _mm_add_epi32(step1[4], step1[5]);
@@ -158,14 +158,14 @@ static INLINE void highbd_idct16x16_10_4col(__m128i *const io /*io[16]*/) {
   __m128i temp[2], sign[2];
 
   // stage 2
-  highbd_partial_butterfly_sse2(io[1], (int)cospi_30_64, (int)cospi_2_64,
-                                &step2[8], &step2[15]);
-  highbd_partial_butterfly_neg_sse2(io[3], (int)cospi_6_64, (int)cospi_26_64,
-                                    &step2[11], &step2[12]);
+  highbd_partial_butterfly_sse2(io[1], cospi_30_64, cospi_2_64, &step2[8],
+                                &step2[15]);
+  highbd_partial_butterfly_neg_sse2(io[3], cospi_6_64, cospi_26_64, &step2[11],
+                                    &step2[12]);
 
   // stage 3
-  highbd_partial_butterfly_sse2(io[2], (int)cospi_28_64, (int)cospi_4_64,
-                                &step1[4], &step1[7]);
+  highbd_partial_butterfly_sse2(io[2], cospi_28_64, cospi_4_64, &step1[4],
+                                &step1[7]);
   step1[8] = step2[8];
   step1[9] = step2[8];
   step1[10] =
@@ -179,13 +179,13 @@ static INLINE void highbd_idct16x16_10_4col(__m128i *const io /*io[16]*/) {
 
   // stage 4
   abs_extend_64bit_sse2(io[0], temp, sign);
-  step2[0] = multiplication_round_shift_sse2(temp, sign, (int)cospi_16_64);
+  step2[0] = multiplication_round_shift_sse2(temp, sign, cospi_16_64);
   step2[1] = step2[0];
   step2[2] = _mm_setzero_si128();
   step2[3] = _mm_setzero_si128();
-  highbd_butterfly_sse2(step1[14], step1[9], (int)cospi_24_64, (int)cospi_8_64,
-                        &step2[9], &step2[14]);
-  highbd_butterfly_sse2(step1[10], step1[13], (int)cospi_8_64, (int)cospi_24_64,
+  highbd_butterfly_sse2(step1[14], step1[9], cospi_24_64, cospi_8_64, &step2[9],
+                        &step2[14]);
+  highbd_butterfly_sse2(step1[10], step1[13], cospi_8_64, cospi_24_64,
                         &step2[13], &step2[10]);
   step2[5] = step1[4];
   step2[6] = step1[7];
