+++ vpx_dsp/x86/highbd_inv_txfm_sse4.h
@@ -67,10 +67,10 @@ static INLINE void highbd_butterfly_cospi16_sse4_1(const __m128i in0,
 
   temp2 = _mm_add_epi32(in0, in1);
   extend_64bit(temp2, temp1);
-  *out0 = multiplication_round_shift_sse4_1(temp1, (int)cospi_16_64);
+  *out0 = multiplication_round_shift_sse4_1(temp1, cospi_16_64);
   temp2 = _mm_sub_epi32(in0, in1);
   extend_64bit(temp2, temp1);
-  *out1 = multiplication_round_shift_sse4_1(temp1, (int)cospi_16_64);
+  *out1 = multiplication_round_shift_sse4_1(temp1, cospi_16_64);
 }
 
 static INLINE void highbd_partial_butterfly_sse4_1(const __m128i in,
